fix: setByUser() is now obeyed.
The bug itself was caused per the "rsc.io/getopt" package since, according to @Vovan-VE at
https://github.com/rsc/getopt/issues/2, it doesn't "inform" the flag package about the
changes that it makes to the argument list, making it null and inutilizing both flag.NArg()
and flag.Visit().
It is far from the state of the art in gambiarra and we could've Lookup()ed the flags
instead of just allocating their name to Lookup() later, so take it as an initial
implementation.